// combined/angular/core/reports/models/report.model.ts
export interface Report {
  cr_status: string;
  lexid: string;
  max_cr_date: string;
  max_pr_date: string;
  max_nr_date: string;
  name_first: string;
  name_last: string;
  pf_status: string;
  pr_status: string;
  record_id: string;
  report_date: string;
  report_type: string;
  status: string;
  unique_id: string;
  watcher: string;
}